model & secuirity Modification

+++ ProjectManagementSystem/Controllers/ItAdminController.cs
@@ -1,8 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
@@ -16,6 +18,7 @@ using ProjectManagementSystem.Models;
 
 namespace ProjectManagementSystem.Controllers
 {
+    [Authorize(Roles = "ItAdmin")]
     public class ItAdminController : Controller
     {
         private readonly UserManager<ApplicationUsers> _usermanager;
@@ -41,11 +44,7 @@ namespace ProjectManagementSystem.Controllers
             return View(alluser);
         }
 
-        // GET: ItAdmin/Details/5
-        public ActionResult Details(int id)
-        {
-            return View();
-        }
+      
 
         // GET: ItAdmin/Create
         public ActionResult AddUser()
@@ -96,23 +95,31 @@ namespace ProjectManagementSystem.Controllers
         public async Task<IActionResult> UpdateUser(string id)
         {
           var user= await _usermanager.FindByIdAsync(id);
-          ViewBag.designation = _role.Roles.Select(r => new SelectListItem { Value = r.Id, Text = r.Name }).ToList();
-          var role = _role.Roles.SingleOrDefault(r => r.Name == user.Designation);
-         
-          
+          if (user.Designation=="ItAdmin")
+          {
+              return ViewBag.massage = "You can not Able to Update Or Delete This User "+user.Name;
+          }
+          else
+          {
+              ViewBag.designation = _role.Roles.Select(r => new SelectListItem { Value = r.Id, Text = r.Name }).ToList();
+              var role = _role.Roles.SingleOrDefault(r => r.Name == user.Designation);
+
+
               UserInfoViewModel userInfo = new UserInfoViewModel
               {
                   Id = user.Id,
                   Name = user.Name,
                   Status = user.Status,
                   Email = user.Email,
-                  Password =user.PasswordHash ,
+                  Password = user.PasswordHash,
                   RoleId = role.Id
-              
-              
-              
+
+
+
               };
               return View(userInfo);
+            }
+         
           
         }
 
@@ -170,25 +177,34 @@ namespace ProjectManagementSystem.Controllers
         public async Task<IActionResult> DeleteUser(string id)
         {
             var user = await _usermanager.FindByIdAsync(id);
-            if (user == null)
-            {
-                ViewBag.errormassage = "uuser Not Found";
-                return View("NotFound");
-            }
-            else
-            {
-                var result = await _usermanager.DeleteAsync(user);
-                if (result.Succeeded)
+            
+                if (user == null)
                 {
-                    return RedirectToAction("Index");
+                    ViewBag.errormassage = "uuser Not Found";
+                    return View("NotFound");
                 }
 
-                foreach (var error in result.Errors)
+                else
                 {
-                    ModelState.AddModelError("", error.Description);
-                }
+                    if (user.Designation == "ItAdmin")
+                    {
+                        return ViewBag.massage = "You can not Able to Update Or Delete This User " + user.Name;
+                    }
+                    else
+                    {
+                    var result = await _usermanager.DeleteAsync(user);
+                    if (result.Succeeded)
+                    {
+                        return RedirectToAction("Index");
+                    }
 
-                return View("Index");
+                    foreach (var error in result.Errors)
+                    {
+                        ModelState.AddModelError("", error.Description);
+                    }
+
+                    return View("Index");
+                }
             }
         }
